adding module_returns. For now, this is how we can talk across modules, since returning values is harder in a loop. This way, we can also set values anywhere and have them accessible everywhere.
also made a number oc hanges to customize rating criteria and rating pages to get things to save and link together.

for now, we're using module_returns to automatically update the factors seen when rating if the user goes back and modifies the customize rating criteria tab. Eventually, we probably want to actually pass this info around, rather than simply triggering a pull from the db. but this was the only way I could figure out to trigger a re-pull